Cuesheet patch from Earl Chew.
When reading the INDEX from the cue sheet, the format MM:SS:FF format
is disallowed if the sample frequency is not a multiple of 75 because
the index would only be approximate. However, 00:00:00 is _exact_
because it denotes the start of the track, so allow it as a special
case.