Add appropriate WAV channel masks for 7 and 8 channel files.
This commit accepts the new default channel masks for 6.1 and 7.1
surround input WAV files, and writes the corresponding masks when
decoding to WAV without a channel mask from the metadata block.

The local copy of the format spec is also updated with the new text
from the flac-website repository.

Patch from Ralph Giles <giles@mozilla.com>