.travis.yml: Generate and validate html documentation.
Install Doxygen when running on the linux environment to test
generation of the API documentation.

Run the included static html documentation as well as the
generated API documentation through xmllint to verify
it conforms to the declared DTD. Since this this will be very
slow or fail relying on network resources, also install
the w3c-sgml library so the xhtml1 DTDs are available locally.

We're checking the source files, not build results in the
particular environment, so it's fine to run this check
on only a subset of builds.

Don't validate on CMake builds since they're not currently
generating API documentation even when Doxygen is available.