include: Escape number signs to suppress warnings
Doxygen sees '#' as a cross-reference, which this is not in this case. This will silence the following warning that pops up frequently when building with docs:

warning: explicit link request to 'defines' could not be resolved

The code shorthand tag was removed in include\FLAC\all.h to remain synonymous with other mentions of #define.