github actions: add a windows cmake target
Make use of the available builders for Microsoft Windows build
converage.

Unfortunately this job reports finding no tests to run, despite
being identical to other builds. Worse, this is not reported as
an error.

Windows doesn't have a package manager like linux and macos, so
until we figure out how to build and cache the ogg dependency
for the windows job, just skip ogg support.