Fix invalid read and remove asserts at new autoc SSE2 routines
Valgrind uncovered a few invalid reads, which this fix should
solve. Also, a few asserts were removed because they do not match
with the actual usage of these functions, i.e., they are quite
often not true. Example: on the very last block of the file, which
can be very small (for example 2 of 3 samples) the chosen autoc
functions with intrinsics are not changed, so on these blocks,
these assert will not hold. They are not neccesary either, as
the line limit++; if(limit < 0) limit = 0; prevents bad things
from happening