Add SSSE3 intrinsic routine for fixed_compute_best_predictor
A new SSSE3 intrinsic routine is added to compute the best
fixed predictor. The existing routine is kept for IA32 systems,
the new routine is only for x86_64 systems as these have 8
extra vector registers, which are needed for the new routine.

According to my testing, the new routine is about 3 times as
fast as the existing routine. This means an overall speedup for
preset -2 of about 30% and for preset -0 of about 15%