Speedup of fixed_compute_best_predictor functions
The code of fixed_compute_best_predictor was presumable once optimized,
but it appears a much more readable form is now much faster:
measurements on a Raspberry Pi 4 (64-bit ARMv8) indicate this code
runs almost twice as fast. One could conjecture this is because the
code was optimized before auto-vectorization of code by compilers
was feasible.