Remove obsolete ci/flac-autotool.sh
I think this is from the jenkins era. For a while we tried to
consolidate continuous-integration build descriptions across
services, but the formats are different enough that it's been
easier to use separate, per-service implementations.

In any case, this isn't used any more so there's no reason to
keep it around.