Fix assertion failure in seek_to_absolute_sample_
On fuzzing the assertion
'decoder->private_->last_frame.header.number_type ==
FLAC__FRAME_NUMBER_TYPE_SAMPLE_NUMBER' failed. This was because it
was possible to get there without having decoded a valid frame yet.

The checks are rearranged such that the code past this assertion is
only reached after it is made sure a valid frame has been decoded.