Do not use main output buffer when outputting padding frames
Currently, when the decoder finds that it is missing some data,
it will fill that up with silence. However, it uses the output
buffer for that, overwriting the last decoded data. Therefore,
more data is dropped than is necessary. This also leads to a
heap use after free when a buffer resize is needed to accomodate
the silence data and the frame data is used for analysis

Credit: Oss-Fuzz
Issue: https://bugs.chromium.org/p/oss-fuzz/issues/detail?id=57354