Hide unreproducible bug that is probably harmless
Issue https://bugs.chromium.org/p/oss-fuzz/issues/detail?id=56395
remains unreproducible to me, and the whole problem doesn't seem
to make any sense. Even downloading the affected build and running
it locally doesn't reproduce the problem.

As this is probably a harmless bug (if it is a bug) that is
triggered by an assert but is handled anyway, I'm hiding the assert
so the handling is fuzzed when the assert is removed, as in
production code.