Only invalidate last seen framesync when actually emptying buffer
The last seen framesync in the bitreader was invalidated when
reading new data in. However, data only leaved the buffer when
more than one word has been fully read. When an invalid frame is
read and reading is aborted within less than a wordlength because
of reaching end-of-stream, the read_from_client had invalidated
the last seen framesync, causing rewinding to the start of the
bitreader. This causes an infinite loop, syncing on the same
pattern every time.