Fix cuesheet parsing code skipping a digit
The code parsing cuepoints of the form MM:SS.SS, which is only
allowed for non-CDDA, had a bug where the first S of the above
template was skipped. That meant that 00:12.34 was parsed as
00:02.34. This is not covered in the test suite, but fuzzing
stumbled on it as 00: as input made the parser skip the string-
terminating nul.

Credit: Oss-Fuzz
Issue: https://bugs.chromium.org/p/oss-fuzz/issues/detail?id=57320