Add C11 threads.h implementation in addition to pthread implementation for multithreaded encoding.
The switching between both APIs is done via macros in a new file
include/share/compat_threads.h, as this seemed simple enough. There are only
slight differences between pthread and C11 threads, in particular the way
attributes are handled for mutex and condition variable (both not used by FLAC),
the signature of the thread function, the return value of trylock, and the error
handling, where pthread uses errno and C11 threads use an explicit return value
enum. The wrapper macros are prefixed with FLAC__ as is common in FLAC code. All
functions are renamed to closely resemble the C11 names instead of the POSIX
names, as this is the more forward-looking approach.

The macro to enable C11 threads is HAVE_C11THREADS. When both HAVE_C11THREADS
and HAVE_PTHREAD are defined, the C11 implementation is preferred.

This does not add any build system support for C11 threads, so the current
default behaviour remains unchanged.