Add Autotools support for C11 threads.
If multithreading is not explicitly disabled, and if pthread is not detected,
enable C11 threads support as a fallback. We still default to pthread as this is
the more conservative and backwards-compatible approach for older POSIX systems.
Using C11 threads might imply using pthreads under the hood, and might require
linking additional pthread libraries, which a simple check for threads.h will
not enable. AX_PTHREAD however does that, and it is thus the safer choice.

configure output is changed to display "C11 threads" or "pthread", instead of
just "yes" when multithreading is enabled.