Add CMake support for C11 threads.
If multithreading is not explicitly disabled, and if pthread is not detected,
enable C11 threads support as a fallback. We still default to pthread as this is
the more conservative and backwards-compatible approach for older POSIX systems.
Using C11 threads might imply using pthreads under the hood, and might require
linking additional pthread libraries, which a simple check for threads.h will
not enable.

In MSVC, threads.h does not compile unless it is in C11 or C17 mode. As
check_include_file() does not honor CMAKE_C_STANDARD or C_STANDARD, this
requires explicitly setting the appropriate verbatim compiler flag in order to
work-around this CMake quirk.

This commit does not enable C11 mode in CMake by default. In order to enable
multithreading with MSVC, it is required to additionally set CMAKE_C_STANDARD=11
or CMAKE_C_STANDARD=17.

Add CI jobs that test with CMAKE_C_STANDARD=11 for MSVC.